Added Ownership for bootcamp

+++ controller/bootcamps.controller.js
@@ -22,20 +22,56 @@ exports.getBootcampDetails = asyncHandler(async (req, res, next) => {
 })
 
 exports.createBootcamp = asyncHandler(async (req, res, next) => {
+
+    //Add user to req bodu
+
+    req.body.user = req.user.id;
+
+    //check for published bootcamp
+    const publishedBootcamp = await Bootcamp.findOne({user: req.user.id});
+
+
+    //If the user is not an admin, they can only add one bootcamp
+    if (publishedBootcamp && req.user.role !== 'admin') {
+        return next(new ErrorResponse(`The user with ID ${req.user.id} has already published a bootcamp`, 400));
+    }
+
+    //Add geocode to body
+    const loc = await geocoder.geocode(req.body.address);
+    req.body.location = {
+        type: 'Point',
+        coordinates: [loc[0].longitude, loc[0].latitude],
+        formattedAddress: loc[0].formattedAddress,
+        street: loc[0].streetName,
+        city: loc[0].city,
+        state: loc[0].stateCode,
+        zipcode: loc[0].zipcode,
+        country: loc[0].country,
+    }
+
+
     const bootcamp = await Bootcamp.create(req.body);
 
     res.status(201).json({success: true, data: bootcamp});
 })
 
 exports.updateBootcamp = asyncHandler(async (req, res, next) => {
-    const bootcamp = await Bootcamp.findByIdAndUpdate(req.params.id, req.body, {
-        new: true,
-        runValidators: true,
-    })
+    let bootcamp = await Bootcamp.findById(req.params.id);
+
     if (!bootcamp) {
         return next(new ErrorResponse(`No Bootcamp found for ${req.params.id}`, 404));
     }
-    res.status(200).json({success: true, data: bootcamp});
+
+    if (bootcamp.user.toString() !== req.user.id && req.user.role !== 'admin') {
+        return next(new ErrorResponse(`User with ID ${req.user.id} is not authorized to update this bootcamp`, 403));
+    }
+
+    bootcamp = await Bootcamp.findOneAndUpdate({"_id": req.params.id}, req.body, {
+        new: true,
+        runValidators: true
+    });
+
+     res.status(200).json({success: true, data: bootcamp});
 
 })
 
@@ -45,8 +81,14 @@ exports.deleteBootcamp = asyncHandler(async (req, res, next) => {
         return next(new ErrorResponse(`No Bootcamp found for ${req.params.id}`, 404));
     }
 
+
+    if (bootcamp.user.toString() !== req.user.id && req.user.role !== 'admin') {
+        return next(new ErrorResponse(`User with ID ${req.user.id} is not authorized to delete this bootcamp`, 403));
+    }
+
+
     const deletedBootcamp = await bootcamp.deleteOne({_id: req.params.id})
-    console.log({deletedBootcamp: deletedBootcamp})
+
     res.status(200).json({success: true, data: {}});
 })
 
@@ -93,6 +135,10 @@ exports.bootcampPhotoUpload = asyncHandler(async (req, res, next) => {
 
     if (!bootcamp) {
         return next(new ErrorResponse(`No Bootcamp found for ${req.params.id}`, 404));
+    }
+
+    if (bootcamp.user.toString() !== req.user.id && req.user.role !== 'admin') {
+        return next(new ErrorResponse(`User with ID ${req.user.id} is not authorized to update this bootcamp`, 403));
     }
 
     if (!req.files) {
